requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.27
botocore==1.38.27
certifi==2025.4.26
click==8.2.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.82.1
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyzbar==0.1.9
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
